# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.TagRequest:
    properties:
      group_id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
  handlers.TagResponse:
    properties:
      created_at:
        type: string
      group_id:
        type: integer
      group_name:
        type: string
      group_slug:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  handlers.TelegramDownloadRequest:
    properties:
      link:
        type: string
      pack_name:
        type: string
    type: object
  handlers.TelegramDownloadResponse:
    properties:
      base_dir:
        type: string
      failed:
        type: integer
      files:
        items:
          $ref: '#/definitions/handlers.TelegramDownloadedFile'
        type: array
      pack_name:
        type: string
      saved:
        type: integer
      title:
        type: string
      total:
        type: integer
    type: object
  handlers.TelegramDownloadedFile:
    properties:
      emoji:
        type: string
      file_id:
        type: string
      file_path:
        type: string
      file_unique_id:
        type: string
      local_path:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for Emoji collection site
  title: Emoji API
  version: 0.1.0
paths:
...
  /api/admin/tags:
    get:
      parameters:
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TagResponse'
            type: array
      summary: List tags
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: tag request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.TagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TagResponse'
      summary: Create tag
      tags:
      - admin
  /api/admin/tags/{id}:
    delete:
      parameters:
      - description: tag id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Delete tag
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: tag id
        in: path
        name: id
        required: true
        type: integer
      - description: tag request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.TagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TagResponse'
      summary: Update tag
      tags:
      - admin
  /api/admin/telegram/download:
    post:
      consumes:
      - application/json
      parameters:
      - description: telegram download
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.TelegramDownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TelegramDownloadResponse'
      summary: Download Telegram sticker pack
      tags:
      - admin
...
swagger: "2.0"